+++ src/main/java/com/lokoko/domain/like/application/service/ReviewLikeService.java
@@ -5,9 +5,9 @@ import com.lokoko.domain.like.domain.repository.ReviewLikeRepository;
 import com.lokoko.domain.review.domain.entity.Review;
 import com.lokoko.domain.review.exception.ReviewNotFoundException;
 import com.lokoko.domain.review.domain.repository.ReviewRepository;
-import com.lokoko.domain.user.entity.User;
+import com.lokoko.domain.user.domain.entity.User;
 import com.lokoko.domain.user.exception.UserNotFoundException;
-import com.lokoko.domain.user.repository.UserRepository;
+import com.lokoko.domain.user.domain.repository.UserRepository;
 import java.util.Optional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
