+++ src/main/java/com/lokoko/domain/like/domain/repository/ReviewLikeRepository.java
@@ -2,7 +2,7 @@ package com.lokoko.domain.like.domain.repository;
 
 import com.lokoko.domain.like.domain.entity.ReviewLike;
 import com.lokoko.domain.review.domain.entity.Review;
-import com.lokoko.domain.user.entity.User;
+import com.lokoko.domain.user.domain.entity.User;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
